Split out the login page functionality into a LoginForm component.
