+++ app/login/page.tsx
@@ -3,37 +3,15 @@
 import styles from '../page.module.css'
 import { signIn } from "next-auth/react";
 import React, { useRef } from "react";
+import LoginForm from '../components/LoginForm'
 
 export default function Login() {
-  const username = useRef("");
-  const pass = useRef("");
-
-  const onSubmit = async () => {
-    const result = await signIn("credentials", {
-      username: username.current,
-      password: pass.current,
-      redirect: true,
-      callbackUrl: "/"
-    });
-  }
 
   return (
     <main className={styles.main}>
-      <h1>Account Login</h1>
-      <div>
-        <div>
-          <p>User Name:</p>
-          <input name="text" placeholder='username' onChange={(e) => (username.current = e.target.value)}></input>
-        </div>
-        <div>
-          <p>Password:</p>
-          <input name="password" type="password" placeholder='password' onChange={(e) => (pass.current = e.target.value)}></input>
-        </div>
-        <button style={{ backgroundColor: "green" }} onClick={onSubmit}>Login</button>
-        <p>If you dont have an account you can <a href="/signup" style={{ color: "blue", textDecoration: "underline" }}>click here </a>to create one.</p>
-        <br></br>
-        <p> test1:123123 and test2:123123 are current test accounts (username:password).</p>
-      </div>
+      <h1 style={{ fontSize: 24, fontWeight: 'bolder' }}>Account Login</h1>
+      <LoginForm/>
+      <p>If you dont have an account you can <a href="/signup" style={{ color: "blue", textDecoration: "underline" }}>click here </a>to create one.</p>
     </main>
   )
 }
